requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
psycopg[binary]>=3.1
python-dotenv>=1.0
pydantic>=2.6
